pyproject.toml
[tool.poetry]
name = "discord-hsreplaynet"
version = "0.1.0"
description = ""
authors = ["Benedict Etzel <benedict@hearthsim.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
sync-roles = "discord_hsreplaynet.bot:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
